# openapi.yaml
openapi: 3.0.0
info:
  title: Hospital Concierge API
  version: 1.0.0
  description: Healthcare SaaS platform API
servers:
  - url: /api
    description: Development server
paths:
...
  /me/visit/book:
    post:
      summary: Book appointment
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BookingRequest'
      responses:
        '200':
          description: Appointment booked
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Appointment'
components:
  schemas:
...
    BookingRequest:
      type: object
      properties:
        providerId:
          type: string
        datetime:
          type: string
        reason:
          type: string
    Appointment:
      type: object
      properties:
        id:
          type: string
        providerId:
          type: string
        datetime:
          type: string
        status:
          type: string